TEAMNADO-2617 - Update Dev Environment to Reflect console.redhat.com Changes (#38)
* Updating app to use the useProxy rather than the old proxy container (no more need to run Docker :)), and updating the menu setup per advice from console dot team to fix the error in which clicking on the manifests link in the side menu, after the app has loaded, results in an error.

* using regex for path replace in getPartialRouteFromPath method
