+++ src/utilities/routeHelpers.ts
@@ -1,0 +1,19 @@
+const getBaseName = (pathname: string): string => {
+  let release = '/';
+  const pathName = pathname.split('/');
+
+  pathName.shift();
+
+  if (pathName[0] === 'beta') {
+    pathName.shift();
+    release = `/beta/`;
+  }
+
+  return `${release}${pathName[0]}/${pathName[1] || ''}`;
+};
+
+const getPartialRouteFromPath = (path: string): string => {
+  return path.replace(/^\/insights\/subscriptions\/manifests\/?/, '/');
+};
+
+export { getPartialRouteFromPath, getBaseName };
